Revert "Adding validators"
This reverts commit 1c38b9dc81e7f92f3515c06243b7536139ef36b8.

+++ backend/src/controllers/user-controller.ts
@@ -1,24 +1,24 @@
-import { Request, Response } from "express";
+import {Request, Response} from "express";
 import { UserModel } from "../models/user-model";
 import * as HttpResponse from "../utils/http-helper";
 import * as UserService from "../services/user-service";
 import { HttpResponseModel } from "../models/http-response-model";
-import { haveAllTheUserParameters } from "../validators/verifyTheParameters";
+import {haveAllTheUserParameters } from "../validators/verifyTheParameters";
 
-export const postUser = async (req: Request, res: Response) => {
+export const postUser  = async (req: Request, res: Response) => {
     const user = req.body as UserModel;
     let httpResponse: HttpResponseModel
 
     const hasAllTheParameters = haveAllTheUserParameters(user);
-
+    
     if (!hasAllTheParameters) {
-        httpResponse = await HttpResponse.badRequest({ message: "Faltam parâmetros" });
+        httpResponse = await HttpResponse.badRequest({ message: "Faltam parâmetros"});
     }
 
     else {
         httpResponse = await UserService.insertUserService(user);
     }
-
+    
     res.status(httpResponse.statusCode).json(httpResponse.body);
 }
 
@@ -30,27 +30,27 @@ export const getAllUsers = async (req: Request, res: Response) => {
 export const getUserByCpf = async (req: Request, res: Response) => {
     const cpf = req.params.cpf.toString();
     let httpResponse = null;
-
-    try {
-        httpResponse = await UserService.getUserByCpfService(cpf);
-        res.status(httpResponse.statusCode).json(httpResponse.body)
-    } 
-    catch (error: any) {
-        res.status(500).json(error.message);
+    
+    //Estou com receio de misturar isso, retornar um HttpResponse no controller, sendo que esse papel, em todos estava sendo unicamente do service!
+    if (!cpf) {
+        httpResponse = await HttpResponse.badRequest({message: "Insira um cpf!"});
     }
+    else {
+        httpResponse = await UserService.getUserByCpfService(cpf);
+    }
+
+    res.status(httpResponse.statusCode).json(httpResponse.body)
 }
 
 export const deleteUserByCpf = async (req: Request, res: Response) => {
     const cpf = req.params.cpf.toString();
     let httpResponse = null;
 
-    try {
+    if (!cpf) {
+        httpResponse = await HttpResponse.badRequest({message: "Insira um cpf!"});
+    }
+    else {
         httpResponse = await UserService.deleteUserByCpfService(cpf);
     }
-    catch (error: any) {
-        httpResponse = await HttpResponse.serverError(error.message);
-    }
-    finally {
-        res.status(httpResponse.statusCode).json(httpResponse.body);
-    }
+    res.status(httpResponse.statusCode).json(httpResponse.body);
 }
